Streaming: Refactor database connection logic to separate module
